[workspaces/kubernetes-native-dxp] automate user.agent profile and big refactor
