+++ workspaces/kubernetes-native-dxp/k8s/remoteappa/deployment.yaml
