--- workspaces/kubernetes-native-dxp/k8s/remoteappa/remote-app-a-configmap.yaml.template
@@ -1,9 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: remote-app-a-configmap
-  labels:
-    dxp: configmap
-data:
-  com.liferay.remote.app.config.factory.configuration.v1.RemoteAppConfiguration-remoteappa.config: |
-    companyId=L"20099"
