+++ workspaces/kubernetes-native-dxp/modules/kubernetes-agent/src/main/java/com/liferay/k8s/agent/configuration/v1/K8sAgentConfiguration.java
@@ -1,0 +1,51 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.k8s.agent.configuration.v1;
+
+import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClassDefinition;
+
+import aQute.bnd.annotation.metatype.Meta;
+import aQute.bnd.annotation.metatype.Meta.Type;
+
+/**
+ * @author Raymond Augé
+ */
+@ExtendedObjectClassDefinition(
+	category = "hidden",
+	scope = ExtendedObjectClassDefinition.Scope.SYSTEM
+)
+@Meta.OCD(
+	factory = true,
+	id = "com.liferay.k8s.agent.configuration.v1.K8sAgentConfiguration",
+	localization = "content/Language", name = "k8s-agent-configuration-name"
+)
+public interface K8sAgentConfiguration {
+
+	@Meta.AD(
+		name = "namespace",
+		description = "namespace-description",
+		type = Type.String
+	)
+	String namespace();
+
+	@Meta.AD(
+		name = "label-selector",
+		description = "label-selector-description",
+		type = Type.String,
+		deflt = "dxp=configmap"
+	)
+	String labelSelector();
+
+}
