+++ workspaces/kubernetes-native-dxp/modules/kubernetes-agent/src/main/resources/com/liferay/k8s/agent/configuration/v1/v1/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
