+++ workspaces/kubernetes-native-dxp/modules/kubernetes-agent/src/main/resources/com/liferay/k8s/agent/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
