--- workspaces/kubernetes-native-dxp/modules/kubernetes-config/bnd.bnd
@@ -1,12 +1,0 @@
-Bundle-Name: Liferay Kubernetes Configuration Agent
-Bundle-SymbolicName: com.liferay.kubernetes.config
-Bundle-Version: 1.0.0
-
-Import-Package: \
-	!android.*,\
-	!javax.annotation.meta.*,\
-	!org.bouncycastle.*,\
-	!org.conscrypt.*,\
-	*
--fixupmessages: Classes found in the wrong directory
--noclassforname: true
