--- workspaces/kubernetes-native-dxp/modules/kubernetes-config/build.gradle
@@ -1,7 +1,0 @@
-dependencies {
-	compileOnly group: "com.liferay.portal", name: "release.dxp.api"
-
-	compileInclude('io.fabric8:kubernetes-client:5.12.2') {
-		exclude group: "org.slf4j", module: "slf4j-api"
-	}
-}
