--- workspaces/kubernetes-native-dxp/modules/oauth2-application-factory-headless-server/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: oauth2-application-factory-headless-server
-Bundle-SymbolicName: com.liferay.oauth2.application.factory.headless.server
-Bundle-Version: 1.0.0
