--- workspaces/kubernetes-native-dxp/modules/oauth2-application-factory-user-agent/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: oauth2-application-factory-user-agent
-Bundle-SymbolicName: com.liferay.oauth2.application.factory.user.agent
-Bundle-Version: 1.0.0
