+++ workspaces/kubernetes-native-dxp/modules/oauth2-application-factory/.gitignore
