+++ workspaces/kubernetes-native-dxp/modules/oauth2-application-factory/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay OAuth2 Application Factory
+Bundle-SymbolicName: com.liferay.oauth2.application.factory
+Bundle-Version: 1.0.0
