+++ workspaces/kubernetes-native-dxp/modules/oauth2-application-factory/build.gradle
@@ -1,0 +1,6 @@
+dependencies {
+	compileOnly group: "com.liferay.portal", name: "release.dxp.api"
+	compileOnly group: "com.liferay", name: "biz.aQute.bnd.annotation", version: "4.2.0.LIFERAY-PATCHED-2"
+
+	compile project(":modules:kubernetes-agent")
+}
