+++ workspaces/kubernetes-native-dxp/modules/oauth2-application-factory/src/main/java/com/liferay/oauth2/application/factory/headless/server/Oauth2ApplicationFactoryHeadlessServer.java
