+++ workspaces/kubernetes-native-dxp/modules/oauth2-application-factory/src/main/java/com/liferay/oauth2/application/factory/user/agent/configuration/v1/OAuth2ApplicationUserAgentConfiguration.java
