--- workspaces/kubernetes-native-dxp/modules/remote-app-config-factory/.gitignore
@@ -1,3 +1,0 @@
-.gradle/
-build/
-target/
