--- workspaces/kubernetes-native-dxp/modules/remote-app-config-factory/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: remote-app-config-factory
-Bundle-SymbolicName: com.liferay.remote.app.config.factory
-Bundle-Version: 1.0.0
