+++ workspaces/kubernetes-native-dxp/modules/remote-app-factory/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Remote App Factory
+Bundle-SymbolicName: com.liferay.remote.app.factory
+Bundle-Version: 1.0.0
