+++ workspaces/kubernetes-native-dxp/modules/remote-app-factory/build.gradle
